Add event listeners in two steps so config event can change them.
Firstly, load all ConfigurationReadyEvent listeners and run them
once the configuration is ready.
Afterwards, load all other event listeners so they can be run.

The goal of this change is to be able to dynamically edit a bundle's
event_listeners attribute after the Configuration is ready.
In other words, for a ConfigurationReadyEvent listener to be able
to add extra listeners that will be run for other events.